auth-service/requirements.txt
fastapi
tortoise-orm
passlib
PyJWT
redis
